requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.35.19
botocore==1.35.19
Brotli==1.1.0
bugsnag==4.7.1
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
django-storages==1.14.4
djangorestframework==3.15.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
ffmpeg==1.4
filelock==3.16.1
fire==0.6.0
fonttools==4.53.0
gunicorn==23.0.0
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
opencv-python-headless==4.10.0.82
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pdf2docx==0.5.8
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytesseract==0.3.10
pytube==15.0.0
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.5
referencing==0.35.1
reportlab==4.2.0
requests==2.32.3
rpds-py==0.18.1
s3transfer==0.10.2
setuptools==70.0.0
simpleaudio==1.0.4
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
termcolor==2.4.0
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.1
urllib3==2.2.1
virtualenv==20.26.5
wcwidth==0.2.13
webencodings==0.5.1
WebOb==1.8.8
wheel==0.43.0
whitenoise==6.7.0
woff2otf==1.0.1
yarg==0.1.9
